Break out source display into own component.
