+++ src/components/SourceDisplay/index.js
@@ -1,0 +1,5 @@
+import ilComponent from './SourceDisplay'
+
+export default Vue => {
+    Vue.component(ilComponent.name, ilComponent);
+}
