+++ src/components/index.js
@@ -5,3 +5,4 @@ export {default as CurationRow} from './CurationRow'
 export {default as AgentSelect} from './AgentSelect'
 export {default as RelationSearch} from './RelationSearch'
 export {default as Relation} from './Relation'
+export {default as SourceDisplay} from './SourceDisplay'
