+++ src/index.js
@@ -20,3 +20,4 @@ export {default as CurationRow} from './components/CurationRow'
 export {default as AgentSelect} from './components/AgentSelect'
 export {default as RelationSearch} from './components/RelationSearch'
 export {default as Relation} from './components/Relation'
+export {default as SourceDisplay} from './components/SourceDisplay'
